requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
pydantic==1.8.1
python-multipart==0.0.5
passlib==1.7.4
SQLAlchemy==1.3.23
async-exit-stack==1.0.1
async-generator==1.10
python-crontab==2.5.1
alembic==1.5.6
